requirements.txt
Django==1.6.2
MySQL-python==1.2.5
django-dajax==0.9.2
django-dajaxice==0.5.5
django-grappelli==2.5.2
django-localeurl==2.0.1
django-money==0.4.1
django-nose==1.2
django-slim==0.7.2
django-suit==0.2.6
django-zurb-foundation==5.1.1
nose==1.3.1
numpy==1.8.0
pandas==0.13.1
py-moneyed==0.5.0
python-dateutil==2.2
pytz==2014.1.1
six==1.4.1
virtualenv==1.11.4
wsgiref==0.1.2
